Minor Improvements. Added Country Partials.

+++ BankTemplateInclDataAccessLayer/ViewModels/CountryBoxViewModel.cs
@@ -7,6 +7,6 @@
         public string CountryCode { get; set; } = null!;
         public ICollection<CustomerViewModel> Customers { get; set; } = new List<CustomerViewModel>();
         public ICollection<AccountViewModel> Accounts { get; set; } = new List<AccountViewModel>();
-        public double BalanceSum { get; set; }
+        public decimal BalanceSum { get; set; }
     }
 }
